requirements.txt
Flask
Gunicorn
requests
requests-oauthlib
rsa
zipp
pandas
Pillow
Flask-DebugToolbar
arabic-reshaper
python-bidi